findTopByAvgScore

+++ rottenpotato/src/main/java/com/redwood/rottenpotato/main/repositories/CriticReviewRepository.java
@@ -26,5 +26,5 @@ public interface CriticReviewRepository extends JpaRepository<CriticReview, Long
 
 
     @Query("select c.itemKey, avg(c.reviewRating) as r, count(c.itemKey) as keycount from CriticReview c where c.reviewRating <>0  group by c.itemKey having  count(c.itemKey)>=:limit order by r desc")
-    List<Object[]> findTopByAvgScore(@Param("limit") int limit);
+    List<Object[]> findTopByAvgScore(@Param("limit") long limit);
 }
